+++ libraries/search_blender.rb
@@ -38,9 +38,7 @@ class ::Chef #nodoc
             else
               results = value.dup
             end
-            output_keys = output_path.split('.')
-            output_entry = output_keys[0...-1].inject(node.override) { |element, key| element[key] }
-            output_entry[output_keys.last] = results
+            Chef::AttributeBlender.blend_attribute_into_node(node, output_path, results)
           end
         end
       end