+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.2:
+
+* Add utility method to blend attribute value into path via
+  `Chef::AttributeBlender.blend_attribute_into_node`
+
 ## 0.2.1:
 
 * Chef 11 compatibility fixes.